requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scikit-learn>=0.24.0
faiss-cpu>=1.7.0
matplotlib>=3.4.0
tqdm>=4.62.0